Add result labels, fix delete, fixes

+++ Project/src/main/java/presentation/controllers/ManageUserController.java
@@ -8,6 +8,9 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.TextField;
+import javafx.scene.control.Label;
+import javafx.scene.paint.Color;
+import javafx.scene.paint.Paint;
 import presentation.userManage.Producer;
 import presentation.userManage.Systemadministrator;
 import presentation.userManage.User;
@@ -18,6 +21,15 @@ public class ManageUserController {
 
     @FXML
     private TextField password;
+
+    @FXML
+    private Label addUserResult;
+
+    @FXML
+    private Label changeUserResult;
+
+    @FXML
+    private Label removeUserResult;
 
     @FXML
     private ComboBox<String> userType;
@@ -36,41 +48,57 @@ public class ManageUserController {
 
     @FXML
     void addUser(ActionEvent event) {
+        boolean success = false;
+        String userUsername = username.getText();
+        String userPassword = DomainFacade.getInstance().generateStrongPasswordHash(password.getText());
+        String userUserType = userType.getValue();
         IUser currentUser = DomainFacade.getInstance().getCurrentUser();
         if (DomainFacade.getInstance().validateUser(currentUser)) {
-            String userUsername = username.getText();
-            String userPassword = DomainFacade.getInstance().generateStrongPasswordHash(password.getText());
-            String userUserType = userType.getValue();
             IUser user = (userUserType.equals("Systemadministrator")) ? new Systemadministrator(userUsername, userPassword) : new Producer(userUsername, userPassword);
-            DomainFacade.getInstance().addUser(user);
+            success = DomainFacade.getInstance().addUser(user);
+        }
+        if (success) {
+            username.setText("");
+            password.setText("");
+            userType.setValue("");
+            addUserResult.setText("Successfully added user: " + userUsername );
+            addUserResult.setTextFill(Color.web("#4BB543"));
+        } else {
+            addUserResult.setText("Error: Something went wrong, try again. \nMake sure that the username does not already exist!");
+            addUserResult.setTextFill(Color.web("#FF0000"));
         }
     }
 
     @FXML
     void deleteUser(ActionEvent event) {
+        String removeUsername = removeSearchUsername.getText();
         IUser currentUser = DomainFacade.getInstance().getCurrentUser();
         boolean success = false;
         if (DomainFacade.getInstance().validateUser(currentUser)) {
-            String removeUsername = removeSearchUsername.getText();
             IUser tempUser = new User(removeUsername);
             IUser removeUser = DomainFacade.getInstance().getUser(tempUser);
-            if (!removeUsername.equals("")) {
-                success = DomainFacade.getInstance().deleteUser(removeUser);
+            if (removeUser != null) {
+                if (!removeUsername.equals("")) {
+                    success = DomainFacade.getInstance().deleteUser(removeUser);
+                }
             }
         }
         if(success) {
-            System.out.println("Nice");
+            removeSearchUsername.setText("");
+            removeUserResult.setText("Successfully removed the user: " +  removeUsername);
+            removeUserResult.setTextFill(Color.web("#4BB543"));
         } else {
-            System.out.println("Sheisse");
+            removeUserResult.setText("Error: Something went wrong, try again. \nMake sure that the user exist!");
+            removeUserResult.setTextFill(Color.web("#FF0000"));
         }
     }
 
     @FXML
     void updateUser(ActionEvent event) {
+        boolean success = false;
         IUser currentUser = DomainFacade.getInstance().getCurrentUser();
         if (DomainFacade.getInstance().validateUser(currentUser)) {
             String searchedUser = editSearchUsername.getText();
-            //Temporary user to fetch the wanted user from the DB.
             IUser tempUser = new User(searchedUser);
             IUser user = DomainFacade.getInstance().getUser(tempUser);
 
@@ -80,7 +108,17 @@ public class ManageUserController {
             if (!changePassword.getText().equals("")) {
                 user.setPassword(DomainFacade.getInstance().generateStrongPasswordHash(changePassword.getText()));
             }
-            DomainFacade.getInstance().editUser(user);
+            success = DomainFacade.getInstance().editUser(user);
+        }
+        if (success) {
+            changeUsername.setText("");
+            changePassword.setText("");
+            editSearchUsername.setText("");
+            changeUserResult.setText("Successfully changed the information");
+            changeUserResult.setTextFill(Color.web("#4BB543"));
+        } else {
+            changeUserResult.setText("Error: Something went wrong, try again. \nMake sure that the username does not already exist!");
+            changeUserResult.setTextFill(Color.web("#FF0000"));
         }
     }
